restaurant-bot-agent/pyproject.toml
[project]
name = "customer-support-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "numpy>=2.4.1",
    "openai-agents[voice]>=0.6.5",
    "python-dotenv>=1.2.1",
    "sounddevice>=0.5.3",
    "streamlit>=1.52.2",
]
